Añadido menú de pausa y probadas el HUD de las habilidades

+++ Assets/Scripts/Item/ItemManager.cs
@@ -16,7 +16,7 @@ public class ItemManager : MonoBehaviour, IControllable
 
     [Header("Item resources")] 
     public float coldownReduced;
-    public float manaStolen;
+    public int manaStolen;
     public float turboRecharged;
 
     private Dictionary<int, List<int>> itemPercentageDictionary;
